+++ controllers/editor/subscription.js
@@ -36,13 +36,15 @@ const deleteSubscriptionById = async (req, res) => {
     if (!subscription)
         return res.status(404).json({ "message": "Sorry! no such subscription found." })
 
-    const path = "public/" + subscription.imageUrl.slice(process.env.BASE_URL.length, subscription.imageUrl.length)
+    if (!subscription.imageUrl) {
+        const path = "public/" + subscription.imageUrl.slice(process.env.BASE_URL.length, subscription.imageUrl.length)
 
-    try {
-        fs.unlinkSync(path);
-        //file removed
-    } catch (err) { console.log(err); }
+        try {
+            fs.unlinkSync(path);
+            //file removed
+        } catch (err) { console.log(err); }
 
+    }
 
     const deleted = await Subscription.destroy({ where: { id: subscriptionId } });
